+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 
+## Unreleased
+
+* Proper handling of changes to INDEX PRIMARY with AUTO_INCREMENT column.
+
+
 ## 1.0.1 (2017-04-02)
 
 * Fix pymysql requirement in package manifest.